# api/static/oma0256_StoreManagerApi_v2_swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "version" : "v2",
    "title" : "Store Manager",
    "description" : "Store Manager is a web application that helps store owners manage sales and product inventory records. This application is meant for use in a single store."
  },
  "servers" : [ {
    "url" : "http://127.0.0.1:5000/",
    "description" : "Development server"
  }, {
    "url" : "https://oma-store-manager-api.herokuapp.com/",
    "description" : "Production server"
  } ],
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  },
  "paths" : {
    "/api/v2/auth/signup" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Add a new store attendant",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "first_name" : {
                    "type" : "string",
                    "example" : "store"
                  },
                  "last_name" : {
                    "type" : "string",
                    "example" : "attendant"
                  },
                  "email" : {
                    "type" : "string",
                    "example" : "attendant@email.com"
                  },
                  "password" : {
                    "type" : "string",
                    "example" : "pass1234"
                  },
                  "confirm_password" : {
                    "type" : "string",
                    "example" : "pass1234"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Successfully add a new store attendant"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized access must be logged in as store owner"
          },
          "403" : {
            "description" : "Forbidden access should be a store owner"
          }
        }
      }
    },
    "/api/v2/auth/login" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "login a user",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "email" : {
                    "type" : "string",
                    "example" : "attendant@email.com"
                  },
                  "password" : {
                    "type" : "string",
                    "example" : "pass1234"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Successfully logged in"
          },
          "400" : {
            "description" : "Bad request"
          }
        }
      }
    },
    "/api/v2/user/{user_id}/toggle-rights" : {
      "get" : {
        "tags" : [ "Authentication" ],
        "summary" : "Toggle's rights of a store attendant",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "in" : "path",
          "name" : "user_id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "description" : "This is the id of a store attendant"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully changed a store attendant role"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized access must be logged in"
          },
          "403" : {
            "description" : "Forbbiden access must be logged in as store owner"
          },
          "404" : {
            "description" : "User with that id does not exist"
          }
        }
      }
    },
...
  }
}